==> notebook/1.ipynb <==
# 1
---
### 오늘 할 일
- 데이터를 뿔리자
- feature 뽑는 코드 작성
- baseline score 내보기

## Positive 기준
- 직접 train에 사용하지 않을 2등 모델의 semantic score가 0.9 이상
- edit distance가 일정 수준 이하
- token 차이가 일정 수준 이하
- tokenizer는 soynlp, sentencepiece 둘다 시도
- 시제가 일정해야 함
- 주체가 일정해야 함


import sys
sys.path.append("/home/angrypark/korean-text-matching-tf/")

from text.tokenizers import JamoTokenizer, TwitterTokenizer, SoyNLPTokenizer, SentencePieceTokenizer

class Config:
    def __init__(self):
        self.sent_piece_model = "/media/scatter/scatterdisk/tokenizer/sent_piece.100K.model"
        self.soynlp_scores = "/media/scatter/scatterdisk/tokenizer/soynlp_scores.sol.100M.txt"
        
config = Config()

tokenizer = SentencePieceTokenizer(config)
# tokenizer = SoyNLPTokenizer(config)

with open("../data/test_queries.txt", "r") as f:
    candidates = [line.strip() for line in f]

tokenized_candidates = list()
for q in candidates:
    tokenized = tokenizer.tokenize(q)
    tokenized_candidates.append(tokenized)

import numpy as np

np.mean([len(q) for q in tokenized_candidates])

from collections import Counter

count = Counter([len(q) for q in tokenized_candidates])

count

import tensorflow as tf

import os

os.path.join

tf.layers.dense()

import editdistance

editdistance.eval("안녕", "안녕하")

import numpy as np

a = np.array([[1,2,3], [2,3,4], [3,4,5]])
a[1:2]

import tensorflow as tf

import editdistance

from collections import Counter

from utils.utils import JamoProcessor

processor = JamoProcessor()

count = list()
with open("../data/small/train.txt", "r") as f:
    for line in f:
        splited = line.strip().split("\t")
        if splited[2]=="1":
            a = processor.word_to_jamo(splited[0]).replace("_", "")
            b = processor.word_to_jamo(splited[1]).replace("_", "")
            count.append(editdistance.eval(a, b))
count = Counter(count)

import seaborn as sns

count

class FeatureExtractor:
    def __init__(self):
        self.tfidf_vectorizer = None
    
    def extract_features(self, A, B, 
                         get_tfidf_similarity=True, 
                         get_edit_distance=True, 
                         get_negative_token_diff=True, ):
        extracted_features = list()
        extracted_features.append()
        
    def token_length(sentence):
        

def feature_extractor(query, reply):
    def token_length(sentence):
        tokenized = tokenizer.tokenize(sentence)
        return len(tokenized)
    
    def sentence_length(sentence):
        return len(sentence)
    
    def tfidf_similarity(sentence):
        
    def edit_distance(query, reply):
        
    def has_negative_token(sentence):
        
    def
