==> notebook/4.ipynb <==
## TFIDF vectorizer 학습

from sklearn.feature_extraction.text import TfidfVectorizer
import sys
sys.path.append("/home/angrypark/korean-text-matching-tf/")

from utils.utils import JamoProcessor
from text.tokenizers import SentencePieceTokenizer

class Config:
    def __init__(self):
        self.sent_piece_model = "/media/scatter/scatterdisk/tokenizer/sent_piece.100K.model"
config = Config()

processor = JamoProcessor()
tokenizer = SentencePieceTokenizer(config)

def my_word_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [word for word in tokenizer.tokenize(raw)]

def my_char_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [processor.word_to_jamo(word) for word in tokenizer.tokenize(raw)]

from sklearn.feature_extraction.text import CountVectorizer

count_word_vectorizer = CountVectorizer(analyzer="word", tokenizer=my_word_tokenizer, max_features=10000)

tfidf_word_vectorizer = TfidfVectorizer(analyzer="word", tokenizer=my_word_tokenizer, max_features=10000)
# tfidf_char_vectorizer = TfidfVectorizer(analyzer="char", tokenizer=my_char_tokenizer)

base_dir = "/media/scatter/scatterdisk/reply_matching_model/normalized/sol.raw_{}.txt"

%%time
train_lines = list()
for i in [2,3,4]:
    with open(base_dir.format(i), "r") as f:
        for line in f:
            train_lines.append(line.split("\t")[1])

%%time
count_word_vectorizer.fit(train_lines)
pickle.dump(count_word_vectorizer, open("../dump/count_word_vectorizer.pkl", "wb"))

import pickle

pickle.dump(count_word_vectorizer, open("../dump/count_word_vectorizer.pkl", "wb"))

%%time
tfidf_word_vectorizer.fit(train_lines)

import pickle

pickle.dump(tfidf_word_vectorizer, open("../dump/tfidf_word_vectorizer_big.pkl", "wb"))
