==> notebook/5.ipynb <==
# 5
- 오늘 할 일
    - 데이터를 뽑기 (기존 데이터 보기 + 데이터 뽑는 과정 보고 추가할 rule 뽑기)
    - 좋은 feature를 많이 뽑기
    - 관련 논문들 읽어보기 + 좋은 feature 많이 뽑기

import pickle
from collections import namedtuple
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
import pandas as pd
import sys
sys.path.append("/home/angrypark/korean-text-matching-tf/")

from utils.utils import JamoProcessor
from text.tokenizers import SentencePieceTokenizer

Config = namedtuple("config", ["sent_piece_model"])
config = Config("/media/scatter/scatterdisk/tokenizer/sent_piece.100K.model")
processor = JamoProcessor()
tokenizer = SentencePieceTokenizer(config)

def my_word_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [word for word in tokenizer.tokenize(raw)]

def my_char_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [processor.word_to_jamo(word) for word in tokenizer.tokenize(raw)]

---
지난 xgboost, catboost가 중요하게 본 feature는 뭐지?

xgb = pickle.load(open("../models/xgb.pkl", "rb"))
cb = pickle.load(open("../models/cb.pkl", "rb"))

from feature_extractor import FeatureExtractor

feature_extractor = FeatureExtractor()

with open("../data/my_data/test.txt", "r") as f:
    test_A, test_B, test_labels = zip(*[line.split("\t") for line in f])
    test_labels = [1 if l.strip()=="1" else 0 for l in test_labels]

%%time
test_set = feature_extractor.extract_features(test_A, test_B)

with open("../data/my_data/train.txt", "r") as f:
    train_A, train_B, train_labels = zip(*[line.strip().split("\t") for line in f])
    train_labels = [1 if l=="1" else 0 for l in train_labels]

%%time
train_set = feature_extractor.extract_features(train_A, train_B)

pd.DataFrame(train_set).to_csv("../data/my_data/train_set.csv", index=False)

with open("../data/my_data/val.txt", "r") as f:
    val_A, val_B, val_labels = zip(*[line.strip().split("\t") for line in f])
    val_labels = [1 if l=="1" else 0 for l in val_labels]

%%time
val_set = feature_extractor.extract_features(val_A, val_B)

pd.DataFrame(val_set).to_csv("../data/my_data/val_set.csv", index=False)

train_set = pd.DataFrame(train_set)

from xgboost import XGBClassifier
xgb = XGBClassifier()

xgb.fit(train_set, train_labels)

xgb.score(pd.DataFrame(test_set), test_labels)

sum(test_labels)

len(test_labels)

sum(train_labels)

len(train_labels)

76486/8276*19292

import pickle

pickle.dump(xgb, open("../models/new_xgb.pkl", "wb"))

xgb.score(pd.DataFrame(train_set), train_labels)
