==> notebook/6_feature_engineering.ipynb <==
# 6. Feature Engineering

import pickle
from collections import namedtuple
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
import pandas as pd
import sys
sys.path.append("/home/angrypark/korean-text-matching-tf/")

from utils.utils import JamoProcessor
from text.tokenizers import SentencePieceTokenizer

Config = namedtuple("config", ["sent_piece_model"])
config = Config("/media/scatter/scatterdisk/tokenizer/sent_piece.100K.model")
processor = JamoProcessor()
tokenizer = SentencePieceTokenizer(config)

def my_word_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [word for word in tokenizer.tokenize(raw)]

def my_char_tokenizer(raw, pos=["Noun", "Alpha", "Verb", "Number"], stopword=[]):
    return [processor.word_to_jamo(word) for word in tokenizer.tokenize(raw)]

from feature_extractor import FeatureExtractor

feature_extractor = FeatureExtractor()

with open("/home/angrypark/paraphrase_detection/data/test_queries.txt", "r") as f:
    _, queries = zip(*[line.strip().split("\t") for line in f])

with open("/home/angrypark/paraphrase_detection/data/test_replies.txt", "r") as f:
    reply_to_idx_dict = dict()
    replies = list()
    for line in f:
        if not line:
            continue
        splits = line.strip().split("\t")
        reply_to_idx_dict[splits[1]] = splits[0]
        replies.append(splits[1])

submission_data = {}
length = len(replies)

from tqdm import tqdm

queries[23:30]

pd.read_csv("../data/my_data/train_set.csv").head()

total = pd.DataFrame()
for query in tqdm(queries):
    A, B = [query]*length, replies
    extracted = feature_extractor.extract_features(A, B)
    extracted = pd.DataFrame(extracted)
    extracted["sentence_A"] = query
    extracted["sentence_B"] = pd.Series(replies)
    total = pd.concat([total, extracted])

A, B = ["사랑해요"] * length, replies

from sklearn.metrics.pairwise import cosine_similarity

word_sim = list()
for a, b in zip(A, B):
    word_sim.append(cosine_similarity(feature_extractor.tfidf_word_vectorizer.transform([a]), 
                                      feature_extractor.tfidf_word_vectorizer.transform([b])))

[(a, b) for a, b in zip(A, B)][0]

feature_extractor.tfidf_word_vectorizer.transform(['사랑해요'])

extracted = feature_extractor.extract_features(A, B)

def model_to_submission(model, name):
    def load_data():
        base_dir = "/home/angrypark/paraphrase_detection/data/small/"
        train_set = pd.read_csv(base_dir + "train_set.csv")
        val_set = pd.read_csv(base_dir + "val_set.csv")
        test_set = pd.read_csv(base_dir + "test_set.csv")
        with open(base_dir + "train.txt", "r") as f:
            _, _, train_labels = zip(*[line.strip().split("\t") for line in f if line])
            train_labels = [1 if l=="1" else 0 for l in train_labels]
        with open(base_dir + "val.txt", "r") as f:
            _, _, val_labels = zip(*[line.strip().split("\t") for line in f if line]) 
            val_labels = [1 if l=="1" else 0 for l in val_labels]
        with open(base_dir + "test.txt", "r") as f:
            _, _, test_labels = zip(*[line.strip().split("\t") for line in f if line])
            test_labels = [1 if l=="1" else 0 for l in test_labels]
            
        return train_set, val_set, test_set, train_labels, val_labels, test_labels
    
    with open("/home/angrypark/paraphrase_detection/data/test_queries.txt", "r") as f:
        _, queries = zip(*[line.strip().split("\t") for line in f])
        
    with open("/home/angrypark/paraphrase_detection/data/test_replies.txt", "r") as f:
        reply_to_idx_dict = dict()
        replies = list()
        for line in f:
            splits = line.strip().split("\t")
            reply_to_idx_dict[splits[1]] = splits[0]
            replies.append(splits[1])
    
    train_set, val_set, test_set, train_labels, val_labels, test_labels = load_data()
    print("Train accuracy : {:.5f}".format(model.score(train_set, train_labels)))
    print("Val   accuracy : {:.5f}".format(model.score(val_set, val_labels)))
    print("Test  accuracy : {:.5f}".format(model.score(test_set, test_labels)))
    
    length = len(replies)
    submit_set = pd.read_csv("../data/submit_set.csv")
    features = list(submit_set.columns).copy()
    features.remove("sentence_A")
    features.remove("sentence_B")
    
    predictions = list()
    for i, query in enumerate(queries):
        data = submit_set.iloc[i*length:(i+1)*length]
        probs = [p[1] for p in model.predict_proba(data[features]).tolist()]
        scores = [(reply, score) for reply, score in zip(replies, probs)]
        predict = sorted(scores, key=lambda x: x[1], reverse=True)[0][0]
        predictions.append(reply_to_idx_dict[predict])
    submission = pd.read_csv("../submission/sample_submission.csv")
    submission["id_script"] = pd.Series(predictions)
    submission.to_csv("../submission/{}.csv".format(name), index=False)
    return submission

xgb_new = pickle.load(open("../models/new_xgb.pkl", "rb"))

_ = model_to_submission(xgb_new, name="xgb_new")

xgb_old = pickle.load(open("../models/xgb.pkl", "rb"))

xgb_submission = model_to_submission(xgb_old, name="xgb_old")

xgb_submission

def make_submission(probs, name):
    """probs를 집어넣으면 바로 submission 파일 생성
    :param probs: {query: [(reply, score), (reply, score)...]} 형태
    :param name: submission의 이름
    """
    with open("/home/angrypark/paraphrase_detection/data/test_replies.txt", "r") as f:
        reply_to_idx_dict = {reply: idx for idx, reply in [line.strip().split("\t") for line in f]}
