==> notebook/9_model_tuning.ipynb <==
# 9. Model Tuning

---
## 오늘 할 일
### Model Tuning
- 기본적인 방법론으로 best model 2~3개 tuning해서 test auc 0.001 이하로 차이나는 best 5 model 뽑기
- 3가지 방법 시도
    - prediction 끼리의 RMSE의 차이를 보면서 가장 맞추는 분야가 다른 2~3개로 ensemble
    - 단순 5개의 average, weighted average 시도
    - 5개 이상의 모델에 대해 meta learner를 학습하여 시도
    
### Feature Engineering
- 더 뽑을 feature 2~3개만 더 찾아보고 결과 확인
---

# Feature Engineering - FuzzyWuzzy

import fuzzywuzzy
import sys
sys.path.append("/Users/angrypark/Desktop/korean-text-matching-tf/")

from utils.utils import JamoProcessor

processor = JamoProcessor()

import pandas as pd
import numpy as np
from sklearn.metrics import roc_auc_score

train_set = pd.read_csv("../data/small/train_set.csv")
val_set = pd.read_csv("../data/small/val_set.csv")
test_set = pd.read_csv("../data/small/test_set.csv")

with open("../data/small/train.txt", "r") as f:
    train_A, train_B, train_labels = zip(*[line.strip().split("\t") for line in f])
    train_labels = [1 if l=="1" else 0 for l in train_labels]
with open("../data/small/val.txt", "r") as f:
    val_A, val_B, val_labels = zip(*[line.strip().split("\t") for line in f])
    val_labels = [1 if l=="1" else 0 for l in val_labels]
with open("../data/small/test.txt", "r") as f:
    test_A, test_B, test_labels = zip(*[line.strip().split("\t") for line in f])
    test_labels = [1 if l=="1" else 0 for l in test_labels]

train_set.head()

from fuzzywuzzy import fuzz

from xgboost import XGBClassifier
import pickle

with open("../data/small/train.txt", "r") as f:
    train_A, train_B, train_labels = zip(*[line.strip().split("\t") for line in f])
    train_labels = [0 if l=="0" else 1 for l in train_labels]
with open("../data/small/val.txt", "r") as f:
    val_A, val_B, val_labels = zip(*[line.strip().split("\t") for line in f])
    val_labels = [0 if l=="0" else 1 for l in val_labels]
with open("../data/small/test.txt", "r") as f:
    test_A, test_B, test_labels = zip(*[line.strip().split("\t") for line in f])
    test_labels = [0 if l=="0" else 1 for l in test_labels]

%%time
xgb = XGBClassifier()
xgb.fit(pd.concat([train_set, val_set]), train_labels + val_labels)
print("Train Accuracy: {:.8f}".format(xgb.score(train_set, train_labels)))
probs = [p[1] for p in model.predict_proba(train_set)]
print("Train ROAUC   : {:.8f}".format(roc_auc_score(train_labels, probs)))
print("Val   Accuracy: {:.8f}".format(xgb.score(val_set, val_labels)))
probs = [p[1] for p in model.predict_proba(val_set)]
print("Val   ROAUC   : {:.8f}".format(roc_auc_score(val_labels, probs)))
print("Test  Accuracy: {:.8f}".format(xgb.score(test_set, test_labels)))
probs = [p[1] for p in model.predict_proba(test_set)]
print("Test  ROAUC   : {:.8f}".format(roc_auc_score(test_labels, probs)))

def model_to_submission(model, name):
    with open("../data/test_queries.txt", "r") as f:
        _, queries = zip(*[line.strip().split("\t") for line in f])
        
    with open("../data/test_replies.txt", "r") as f:
        reply_to_idx_dict = dict()
        replies = list()
        for line in f:
            splits = line.strip().split("\t")
            reply_to_idx_dict[splits[1]] = splits[0]
            replies.append(splits[1])
    
    print("Train accuracy : {:.8f}".format(model.score(train_set, train_labels)))
    probs = [p[1] for p in model.predict_proba(train_set)]
    print("Train ROAUC    : {:.8f}".format(roc_auc_score(train_labels, probs)))
    print("Val   accuracy : {:.8f}".format(model.score(val_set, val_labels)))
    probs = [p[1] for p in model.predict_proba(val_set)]
    print("Val   ROAUC    : {:.8f}".format(roc_auc_score(val_labels, probs)))
    print("Test  accuracy : {:.8f}".format(model.score(test_set, test_labels)))
    probs = [p[1] for p in model.predict_proba(test_set)]
    print("Test  ROAUC    : {:.8f}".format(roc_auc_score(test_labels, probs)))
    
    length = len(replies)
    submit_set = pd.read_csv("../data/submit_set.csv")
    features = list(submit_set.columns).copy()
    features.remove("sentence_A")
    features.remove("sentence_B")
    
    predictions = list()
    for i, query in enumerate(queries):
        data = submit_set.iloc[i*length:(i+1)*length]
        probs = [p[1] for p in model.predict_proba(data[features]).tolist()]
        scores = [(reply, score) for reply, score in zip(replies, probs)]
        predict = sorted(scores, key=lambda x: x[1], reverse=True)[0][0]
        predictions.append(reply_to_idx_dict[predict])
    submission = pd.read_csv("../submission/sample_submission.csv")
    submission["id_script"] = pd.Series(predictions)
    submission.to_csv("../submission/{}.csv".format(name), index=False)
    return submission

_ = model_to_submission(xgb, name="8")

fuzz_simple_ratio = [fuzz.ratio(processor.word_to_jamo(a), processor.word_to_jamo(b)) 
                           for a, b in zip(train_A, train_B)] 
token_sort_ratio = [fuzz.token_sort_ratio(a, b) for a, b in zip(train_A, train_B)]
train_set["fuzz_simple_ratio"] = fuzz_simple_ratio
train_set["fuzz_token_sort_ratio"] = token_sort_ratio

fuzz_simple_ratio = [fuzz.ratio(processor.word_to_jamo(a), processor.word_to_jamo(b)) 
                           for a, b in zip(val_A, val_B)] 
token_sort_ratio = [fuzz.token_sort_ratio(a, b) for a, b in zip(val_A, val_B)]
val_set["fuzz_simple_ratio"] = fuzz_simple_ratio
val_set["fuzz_token_sort_ratio"] = token_sort_ratio

fuzz_simple_ratio = [fuzz.ratio(processor.word_to_jamo(a), processor.word_to_jamo(b)) 
                           for a, b in zip(test_A, test_B)] 
token_sort_ratio = [fuzz.token_sort_ratio(a, b) for a, b in zip(test_A, test_B)]
test_set["fuzz_simple_ratio"] = fuzz_simple_ratio
test_set["fuzz_token_sort_ratio"] = token_sort_ratio

%%time
xgb = XGBClassifier()
xgb.fit(pd.concat([train_set, val_set]), train_labels + val_labels)
print("Train Accuracy: {:.8f}".format(xgb.score(train_set, train_labels)))
probs = [p[1] for p in xgb.predict_proba(train_set)]
print("Train ROAUC   : {:.8f}".format(roc_auc_score(train_labels, probs)))
print("Val   Accuracy: {:.8f}".format(xgb.score(val_set, val_labels)))
probs = [p[1] for p in xgb.predict_proba(val_set)]
print("Val   ROAUC   : {:.8f}".format(roc_auc_score(val_labels, probs)))
print("Test  Accuracy: {:.8f}".format(xgb.score(test_set, test_labels)))
probs = [p[1] for p in xgb.predict_proba(test_set)]
print("Test  ROAUC   : {:.8f}".format(roc_auc_score(test_labels, probs)))

submit_set = pd.read_csv("../data/submit_set.csv")
with open("../data/submit_lines.txt", "r") as f:
    submit_A, submit_B = zip(*[line.strip().split("\t") for line in f])

fuzz_simple_ratio = [fuzz.ratio(processor.word_to_jamo(a), processor.word_to_jamo(b)) 
                           for a, b in zip(submit_A, submit_B)] 
token_sort_ratio = [fuzz.token_sort_ratio(a, b) for a, b in zip(submit_A, submit_B)]
submit_set["fuzz_simple_ratio"] = fuzz_simple_ratio
submit_set["fuzz_token_sort_ratio"] = token_sort_ratio

train_set.to_csv("../data/small/train_set.csv", index=False)
val_set.to_csv("../data/small/val_set.csv", index=False)
test_set.to_csv("../data/small/test_set.csv", index=False)
submit_set.to_csv("../data/submit_set.csv", index=False)

_ = model_to_submission(xgb, name="9")

feature_importance = sorted([(feat, imp) for feat, imp in zip(list(train_set.columns), xgb.feature_importances_)], key=lambda x: x[1], reverse=True)
for feat, imp in feature_importance:
    print("{} : {}".format(feat, imp))

Train Accuracy: 0.85917391
Train ROAUC   : 0.93523988
Val   Accuracy: 0.74549893
Val   ROAUC   : 0.83392629
Test  Accuracy: 0.74257298
Test  ROAUC   : 0.82238192

from catboost import CatBoostClassifier

%%time
cb = CatBoostClassifier(verbose=False)
cb.fit(pd.concat([train_set, val_set]), train_labels + val_labels)
print("Train Accuracy: {:.8f}".format(cb.score(train_set, train_labels)))
probs = [p[1] for p in cb.predict_proba(train_set)]
print("Train ROAUC   : {:.8f}".format(roc_auc_score(train_labels, probs)))
print("Val   Accuracy: {:.8f}".format(cb.score(val_set, val_labels)))
probs = [p[1] for p in cb.predict_proba(val_set)]
print("Val   ROAUC   : {:.8f}".format(roc_auc_score(val_labels, probs)))
print("Test  Accuracy: {:.8f}".format(cb.score(test_set, test_labels)))
probs = [p[1] for p in cb.predict_proba(test_set)]
print("Test  ROAUC   : {:.8f}".format(roc_auc_score(test_labels, probs)))
